+++ src/components/impact/ImpactCounter.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import React from 'react'
-import impactData from '../../data/impact.json'
+import impactData from '@/data/impact.json'
 
 // Only show if we have valid data (CI Trigger)
 const showStats =